+++ app/Http/Controllers/InvoiceController.php
@@ -13,6 +13,7 @@ use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class InvoiceController extends Controller
@@ -157,6 +158,10 @@ class InvoiceController extends Controller
      */
     public function destroy(Invoice $invoice): JsonResponse
     {
+        if(!Auth::user()->tokenCan('delete')){
+            return response()->json([], 401);
+        }
+
         $wasDeleted = $invoice->delete();
 
         if ($wasDeleted) {
